ChoiceList refactored
